// client/src/types.ts
export type CustomerCountry = 'Estonia' | 'Finland' | 'Norway' | 'Chile' | 'Canada';

export interface DatabaseCustomer {
    id: string;
    first_name: string;
    last_name: string;
    country: CustomerCountry;
    created_at: Date;
    updated_at: Date;
};

export type LeaderboardCustomer = {
    name: string;
    country: CustomerCountry;
    totalBets: number;
    winRate: number;
    profit: number;
}; 